Demonstrate difference between base64{,ct} WRT the last character
The base64 crate notices when the last character contains non-zero bits
that are lost due to there not being an output byte for them to go into.

The base64ct crate does not.

A selling point of base64ct is its strictness. This is so not-strict
that I believe it's right to consider it a bug.